feat: add the user object infos to card
Add city, state and description to User interface in Auth provider.
Add city, state and description to the user configuration modal.
Add a condition to configuration modal submit function to check if the form camp is
filled. If is not filled, only closes the modal and doesn't make the
request. Necessary to not overwrite all user data.